Move global store
