+++ composables/stores/globals.ts
